* Move IdxUrlHelper to Helpers folder. * Fix OKTA-388344 (Normalize Issuer URL)
